Fix go.mod to have the correct Go version
As of recent Go versions, specifying the go directive without a patch
(e.g. 1.22) is no longer supported. Because go tries to download a
matching toolchain if you're not already using one, it will try to
download go toolchain version 1.22, which doesn't exist (1.22.0 does).

Fix the go.mod version to specify the full version with patch, 1.22.0.

Signed-off-by: Kevin Parsons <kevpar@microsoft.com>