Revert "Merged PR 11595595: [release/internal/0.128] Fix mem leak in newHcsTask()
newHcsTask() starts a goroutine for hyperv containers to enable hcsTask and hcsExec
resource cleanup as expected. But when containers exit before the uvm in the normal
case, this goroutine is leaked as it is waiting on a block wait on the UVM.
This commit ensures that we wait for container exit as well so that we do not leak the routine.

**Testing/Validation:**

Validated via debugger that we are terminating the goroutine and also freeing
up resources.
Tested with multiple container recreates on the same UVM and verified with pprof.

**Note**: This is an interim fix. The cleaner approach needs more code changes in well known code paths and more validation to ensure we are not breaking any existing flow. Since timeline to get this fix in is sensitive, proposing this fix for next patch release of release/1.28 so that the risk is very low. We will be revisiting this again and have a cleaner approach on github/hcsshim/main and future minor version pkg releases.

----
#### AI description  (iteration 1)
#### PR Classification
Bug fix

#### PR Summary
This pull request addresses a memory leak issue in the `newHcsTask` function by ensuring proper cleanup of resources when a container or its UVM exits.
- `internal/uvm/wait.go`: Added `WaitForUvmOrContainerExit` function to handle container and UVM exit.
- `cmd/containerd-shim-runhcs-v1/task_hcs.go`: Updated to use `WaitForUvmOrContainerExit` for proper resource cleanup and renamed related functions for clarity.

(cherry picked from commit 7e3f6c85966129ec3ef709b5578781bb3f7af7c9)
Signed-off-by: Kirtana Ashok <kiashok@microsoft.com>"